Moved sharedUtils.ts
