+++ src/routes/(app)/booking/admin/+page.server.ts
@@ -1,7 +1,7 @@
 import apiNames from "$lib/utils/apiNames";
 import { authorize } from "$lib/utils/authorization";
 import type { PageServerLoad } from "./$types";
-import { actions, getAllBookingRequestsWeekly } from "./sharedUtils";
+import { actions, getAllBookingRequestsWeekly } from "../sharedUtils";
 
 export const load: PageServerLoad = async ({ locals }) => {
   const { prisma, user } = locals;
